Fix pyplotgen timeheight cmaps (#1212)
* Fixed time-height plot colormaps
CLUBB ticket #1199

Added a "quick" fix for time-height plot colormaps when variables are not correlations (i.e. values in [-1;1])
Updated VariableGroupCorrelations to include all correlations that are usually output in CLUBB.

* Fixed time-height plot colormaps
CLUBB ticket #1199

Fixed distinction for colormaps in time-height plots and added tick formatter.
We now check the labels for the string "corr" to find correlation variables.
Otherwise we just use one-sided or two-sided normed colormaps.