Clubb silhs gg (#1239)
* I added new subroutine pdf_hydromet_microphys_prep to encompass
the code that calls the hydrometeor PDF and then ultimately
handles the setup of SILHS sample points when SILHS is in use.

* Placeholder subroutine for SILHS grid generalization test.

* I have now connected the new subroutine to run_clubb. The results
for all cases match bit-for-bit to the previous revision.

* Cleaning up from the refactoring of the hydrometeor PDF and SILHS
code into a wrapper subroutine.

* I'm moving pdf_hydromet_microphys_prep to its own file to avoid
circular dependency once it finally needs to be called from the
generalized grid test code soon.

* Added some code to the SILHS generalized grid test.

* Changes to make this a fully functioning grid generalization test
for the part of the microphysics code that encompasses
setup_pdf_parameters and hydrometeor_mixed_moments.

* I have completed the code for the SILHS generalized grid test.

* I generalized precipitation fraction for ascending vs. descending grid.

* Fixing a bug that I noticed in the precipitation fraction calculation
when the component precip frac calculation was set to option 1
(which is currently not the default).

* The code should now be successfully grid-generalized for SILHS.
The ascending vs descending grid test should work now for all cases.

* Updating the core flags .json file to change some of the SILHS flags.

* I moved the call to advance_clubb_core_api (the default run option)
to the top of the "if" block, ahead of the grid generalization test
option that is now in the "else" block.

* Adding a new, hardwired constant flag called l_ascending_grid and
setting it to true in clubb_driver.F90.

* Keeping lines under 100 characters in length.

* Making it so the ascending vs. descending grid tests pass for the
l_lh_random_start flag.